new version
added a hash table for storing the heuristic of a state so it will not
calculate equal states heuristic twice.
changed the state array to byte[][]

+++ src/GameLogic.java
@@ -22,7 +22,7 @@ public class GameLogic implements IGameLogic {
 		this.y = y;
 		this.playerID = playerID;
 		pControl = new PatternControl();
-		currentState = new State(new int[x][y], new int[x]);
+		currentState = new State(new byte[x][y], new int[x]);
 		transitionController = new TransitionController();
 		tree = new GameTree(playerID);
 	}
